requirements.txt
google-adk>=1.25.0
a2a-sdk[http-server]>=0.3.26,<1.0.0
httpx>=0.28.0
python-dotenv>=1.0.0
litellm==1.83.7
uvicorn>=0.41.0
